Add run option for a code reference
When the run option is given, we no longer exec the script; we just call
the code ref.

One way to use this with Oddmsue -- a CGI script that can be used in a
mod_perl environment -- is as follows:

plugin CGI => {
  route => '/wiki',
  script => 'wiki.pl',
  run => \&OddMuse::DoWikiRequest,
  before => sub {
    $OddMuse::RunCGI = 0;
    $OddMuse::DataDir = '/tmp/oddmuse';
    require 'wiki.pl' unless defined &OddMuse::DoWikiRequest;
  },
  env => {},
  errlog => 'wiki.log', # path to where STDERR from cgi script goes
};

We load the script in the before hook if it hasn't been loaded before
and then we just use the entry point for the new run option.