Make sure the status code is actually used
If the CGI script uses the -nph option, $headers will start with a line
such as "HTTP 404 NOT FOUND" and it will contain a Status header saying
"404 NOT FOUND". Just calling $c->res->parse($headers); doesn't handle
this, apparently. We're always getting a 200 or 302 code; the 302 code
is what we get based on the Location header, I guess. We need to
manually set the result code.