Make sure the status code is used always
If the CGI script gives a CGI response (so no "non-parsed headers),
$headers will contain a line such as "Status_ 304 NOT MODIFIED". Just
calling $c->res->parse($headers); doesn't handle this, apparently. We're
always getting a 200 or 302 code. We need to manually set the result
code. We also cannot change the order of parsing because we cannot
change the code after parsing.