Testing and workaround for the NPH setup added
As it turns out, header parsing actually works fine in the NPH case, if
and only if the Status Line actually contains a version as specified.
Unfortunately, the CGI module itself doesn't emit a version if the
version is not mentioned in the SERVER_PROTOCOL environment variable.
And Mojolicious::Plugin::CGI doesn't do that. This is mandatory,
however. The reference for SERVER_PROTOCOL is section 4.1.16 of RFC
3875. Sadly, it's not immediately obvious, which version we should
specify. Always 1.1?

Thus, the workaround: If and only if the Status Line does not contain a
version, we attempt to extract the code anyway and set it ourselves.

Here's how to reproduce the problem:

$ perl -e 'use CGI qw/-nph/; $q = new CGI; print $q->header(-status => "404 NOT MODIFIED")'
HTTP/1.0 404 NOT MODIFIED
Server: cmdline
Status: 404 NOT MODIFIED
Date: Sun, 11 Oct 2015 15:24:33 GMT
Content-Type: text/html; charset=ISO-8859-1

$ export SERVER_PROTOCOL=HTTP
$ perl -e 'use CGI qw/-nph/; $q = new CGI; print $q->header(-status => "404 NOT MODIFIED")'
HTTP 404 NOT MODIFIED
Server: cmdline
Status: 404 NOT MODIFIED
Date: Sun, 11 Oct 2015 15:24:48 GMT
Content-Type: text/html; charset=ISO-8859-1